Moved event loop into canvas
